Добавил текстовые и цифровые константы, перенес функцию валидации в util.js
