+++ js/util.js
@@ -20,3 +20,16 @@ export const getLengthCompare = (str, maxLength) => {
 
 export const isEscapeKey = (evt) => evt.key === 'Escape';
 
+export const validHashTag = (value) => {
+  const hashtagsList = value.split(/(?=#)/g);
+  const hashtagCondition = /^#[a-zа-я0-9]{1,19}$/i;
+  if(hashtagsList[0] === ''){
+    return true;
+  }
+  for (let i = 0; i < hashtagsList.length; i++){
+    if(!hashtagCondition.test(hashtagsList[i])){
+      return false;
+    }
+  }
+  return true;
+};
